requirements.txt
# Core dependencies for the Chameleon MCP Workflow project

# MCP Server dependencies
mcp[cli]>=0.1.0

# Streamlit for client UI
streamlit>=1.28.0
streamlit-autorefresh>=1.0.0

# Graph visualization
graphviz>=0.20.0

# Workflow engine dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Async support
httpx>=0.25.0
aiohttp>=3.9.0

# JWT authentication (Phase 2)
PyJWT>=2.8.0

# Configuration and environment
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# Data handling
pydantic-settings>=2.0.0
python-dateutil>=2.8.0

# Database
sqlalchemy>=2.0.0
teradatasqlalchemy>=17.0.0

# Template rendering
jinja2>=3.1.0

# Configuration file support
PyYAML>=6.0.1

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality (optional, for development)
ruff>=0.1.0
black>=23.0.0
mypy>=1.5.0

# AI Integration support
# Note: GitHub Copilot, Claude, and Antigravity are used during development
# Antigravity is part of Python standard library (easter egg)